# docs/api/control-v1.openapi.yaml
openapi: 3.1.0
info:
  title: ya-router Control API
  version: 1.0.0
  description: 'Isolated management-plane contract for ya-routerd. The Control API
    does

    not accept YA_ROUTER_API_KEY and is not served from the OpenAI-compatible

    data-plane listener.

    '
servers:
- url: http://unix/control/v1
  description: Local Unix socket transport (socket path is deployment-specific)
- url: https://127.0.0.1:7443/control/v1
  description: Optional loopback TLS transport
paths:
...
  /operations:
    get:
      operationId: listControlOperations
      summary: List asynchronous operations
      security: *id001
      parameters:
      - $ref: '#/components/parameters/RequestID'
      - $ref: '#/components/parameters/ClientVersion'
      responses:
        '200':
          description: Successful redacted read
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationCollection'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '426':
          $ref: '#/components/responses/UpgradeRequired'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      operationId: createControlOperation
      summary: Create a persistent provider-neutral operation
      security: &id002
      - localUnixSocket: []
      - controlBearer: []
      - mutualTLS: []
      parameters:
      - &id003
        $ref: '#/components/parameters/RequestID'
      - &id004
        $ref: '#/components/parameters/ClientVersion'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOperationRequest'
      responses:
        '200':
          description: Existing idempotent operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
        '202':
          description: Operation accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
        '409':
          $ref: '#/components/responses/Conflict'
        '400': &id005
          $ref: '#/components/responses/BadRequest'
        '401': &id006
          $ref: '#/components/responses/Unauthorized'
        '403': &id007
          $ref: '#/components/responses/Forbidden'
        '426': &id008
          $ref: '#/components/responses/UpgradeRequired'
        '500': &id009
          $ref: '#/components/responses/InternalError'
...
  /operations/{operation}:
    parameters:
    - $ref: '#/components/parameters/OperationID'
    get:
      operationId: getControlOperation
      summary: Read one visible operation
      security: *id002
      parameters: &id010
      - *id003
      - *id004
      responses:
        '200':
          description: Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
        '404':
          $ref: '#/components/responses/NotFound'
        '400': *id005
        '401': *id006
        '403': *id007
        '426': *id008
        '500': *id009
    delete:
      operationId: cancelControlOperation
      summary: Cancel an owned cancelable operation
      security: *id002
      parameters: *id010
      responses:
        '200':
          description: Cancelled operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '400': *id005
        '401': *id006
        '403': *id007
        '426': *id008
        '500': *id009
...
  /auth-sessions:
    post:
      operationId: createAuthSession
      summary: Create a provider-neutral persistent authentication session
      description: Starts a durable provider-specific authentication workflow.
        API-key values are supplied only through the write-only secret resource;
        raw secrets are rejected by this endpoint.
      security: *id002
      parameters:
      - *id003
      - *id004
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAuthSessionRequest'
      responses:
        '200':
          description: Existing idempotent auth session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
        '202':
          description: Auth session accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
        '409':
          $ref: '#/components/responses/Conflict'
        '400': *id005
        '401': *id006
        '403': *id007
        '426': *id008
        '500': *id009
...
components:
  securitySchemes:
    localUnixSocket:
      type: apiKey
      in: header
      name: X-Local-Unix-Socket-Identity
      description: 'Documentation marker only. Local identity is established by Unix
        socket

        ownership and mode; clients do not send this header.

        '
    controlBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-control-token
      description: 'Independent viewer/operator/admin control token. It must differ
        from

        YA_ROUTER_API_KEY and is allowed only for explicitly configured loopback

        TLS administration.

        '
    mutualTLS:
      type: mutualTLS
      description: 'Required for non-loopback remote administration. Verified URI
        SAN,

        full subject, or Common Name values map to viewer/operator/admin roles.

        '
  parameters:
    RequestID:
      name: X-Request-ID
      in: header
      required: false
      schema:
        type: string
        minLength: 8
        maxLength: 128
      description: Optional safe caller correlation ID. The service generates one
        otherwise.
    ClientVersion:
      name: X-YA-Client-Version
      in: header
      required: false
      schema:
        type: string
        examples:
        - 1.1.0
        - 1.0.0
      description: ya client semantic version used for N/N-1 negotiation.
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 200
      description: Persistent owner-scoped operation creation key.
    OperationID:
      name: operation
      in: path
      required: true
      schema:
        type: string
        pattern: ^op_[a-f0-9]{32}$
  headers:
    RequestID:
      schema:
        type: string
      description: Correlation ID included in logs, audit, and typed errors.
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required:
      - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
    ErrorBody:
      type: object
      additionalProperties: false
      required:
      - code
      - message
      - retryable
      - request_id
      properties:
        code:
          type: string
        message:
          type: string
        retryable:
          type: boolean
        request_id:
          type: string
        details:
          type: object
          additionalProperties: true
...
    OperationCollection:
      type: object
      required:
      - data
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Operation'
...
    OperationFailure:
      type: object
      additionalProperties: false
      required:
      - code
      - message
      - retryable
      properties:
        code:
          type: string
        message:
          type: string
        retryable:
          type: boolean
    Operation:
      type: object
      additionalProperties: false
      required:
      - id
      - kind
      - owner
      - state
      - progress
      - cancelable
      - recovery_policy
      - created_at
      - updated_at
      - expires_at
      - sequence
      properties:
        id:
          type: string
        kind:
          type: string
        target:
          type: string
        owner:
          type: string
        state:
          type: string
          enum:
          - pending
          - running
          - waiting_for_user
          - succeeded
          - failed
          - cancelled
          - expired
        progress:
          type: integer
          minimum: 0
          maximum: 100
        cancelable:
          type: boolean
        recovery_policy:
          type: string
          enum:
          - fail
          - expire
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        sequence:
          type: integer
          format: uint64
        failure:
          $ref: '#/components/schemas/OperationFailure'
        metadata:
          type: object
          additionalProperties:
            type: string
        result:
          type: object
          additionalProperties:
            type: string
    CreateOperationRequest:
      type: object
      additionalProperties: false
      required:
      - kind
      properties:
        kind:
          type: string
          enum:
          - config_apply
          - config_validate
          - credential_verify
          - model_refresh
        target:
          type: string
          maxLength: 256
        cancelable:
          type: boolean
          default: true
        expires_in_seconds:
          type: integer
          minimum: 30
          maximum: 3600
        recovery_policy:
          type: string
          enum:
          - fail
          - expire
    CreateAuthSessionRequest:
      type: object
      additionalProperties: false
      required:
      - provider_id
      - method
      properties:
        provider_id:
          type: string
          enum:
          - copilot
          - codex
          - kilo
        account_id:
          type: string
        method:
          type: string
          enum:
          - device_code
          - api_key
          - manual_token_recovery
          - anonymous
        expires_in_seconds:
          type: integer
          minimum: 30
          maximum: 3600
...
  responses:
    Unauthorized:
      description: Missing, invalid, or data-plane-only credential
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: Authenticated identity lacks the required role
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    InternalError:
      description: Sanitized internal failure
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    BadRequest:
      description: Invalid request or cursor
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    UpgradeRequired:
      description: Client version is outside the supported N/N-1 window
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: State or idempotency conflict
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Operation or auth session was not found
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'